+++ .pre-commit-config.yaml
@@ -19,7 +19,7 @@ repos:
       hooks:
           - id: blacken-docs
     - repo: https://github.com/PyCQA/isort
-      rev: 5.10.1
+      rev: 5.11.5
       hooks:
           - id: isort
     - repo: https://github.com/asottile/yesqa
